refactor theme components into dedicated subfolder
