+++ src/components/theme/styles/animated-theme-toggler.css
